Revert "[OSS][POC] Add GrapheneHealthChangedEvent + add _useAssetLiveDataProv…"
This reverts commit 235ccd4d99b1e43e9ac82dcf4855a4003e57beae.
