+++ js_modules/dagster-ui/packages/ui-core/client.json
@@ -123,8 +123,8 @@
   "CapturedLogsSubscription": "fa5e55b59e9d8632ae71a8387c54230ba71e6f57849a974225ba039808acfa93",
   "CapturedLogsMetadataQuery": "b59ada7585593473002a7b044f09daa85f160445cbc9a4e8ffe0b46d51875cb1",
   "CapturedLogsQuery": "872b617b4f33ee5f6feeba9a4c76ec986fca357695a114e3d7b63172e4600b57",
-  "PipelineRunLogsSubscription": "929d969805de91b8890960433258984bbd32d2c08dbca55583da64c309fa3127",
-  "RunLogsQuery": "811ef63a737599640cc1677f4a6c01c5193e5651e0b641680847abe613296deb",
+  "PipelineRunLogsSubscription": "564c771292632773b15107100b12817e8058821400295dda573c7fa6ee87ce69",
+  "RunLogsQuery": "c1d1c5ceeddbe3ab4283b5ade99eaece9dea9eb4e85d5f5db8b32ecb412b691a",
   "QueuedRunCriteriaQuery": "f925e1576792948317528a64856c3c020591feefac39f1aaebb3092d0fd3cd0b",
   "QueueDaemonStatusQuery": "aa51c596ee907346e60e2fe173bba10ae2ead067d45109225a2cd400a2278841",
   "PipelineEnvironmentQuery": "3b668b028997fb35b17b4d8a90a18b78dd8a70910f2c12aac63065c0584e3a10",
