+++ js_modules/dagster-ui/packages/ui-core/src/runs/LogsRowStructuredContent.tsx
@@ -133,7 +133,6 @@ export const LogsRowStructuredContent = ({node, metadata}: IStructuredContentPro
         </DefaultContent>
       );
     case 'FailedToMaterializeEvent':
-    case 'HealthChangedEvent':
     case 'MaterializationEvent':
       return (
         <AssetMetadataContent
