+++ js_modules/dagster-ui/packages/ui-core/src/runs/types/RunFragments.types.ts
@@ -1483,167 +1483,6 @@ export type RunDagsterRunEventFragment_HandledOutputEvent = {
   >;
 };
 
-export type RunDagsterRunEventFragment_HealthChangedEvent = {
-  __typename: 'HealthChangedEvent';
-  message: string;
-  timestamp: string;
-  level: Types.LogLevel;
-  stepKey: string | null;
-  eventType: Types.DagsterEventType | null;
-  label: string | null;
-  description: string | null;
-  partition: string | null;
-  metadataEntries: Array<
-    | {
-        __typename: 'AssetMetadataEntry';
-        label: string;
-        description: string | null;
-        assetKey: {__typename: 'AssetKey'; path: Array<string>};
-      }
-    | {
-        __typename: 'BoolMetadataEntry';
-        boolValue: boolean | null;
-        label: string;
-        description: string | null;
-      }
-    | {
-        __typename: 'CodeReferencesMetadataEntry';
-        label: string;
-        description: string | null;
-        codeReferences: Array<
-          | {
-              __typename: 'LocalFileCodeReference';
-              filePath: string;
-              lineNumber: number | null;
-              label: string | null;
-            }
-          | {__typename: 'UrlCodeReference'; url: string; label: string | null}
-        >;
-      }
-    | {
-        __typename: 'FloatMetadataEntry';
-        floatValue: number | null;
-        label: string;
-        description: string | null;
-      }
-    | {
-        __typename: 'IntMetadataEntry';
-        intValue: number | null;
-        intRepr: string;
-        label: string;
-        description: string | null;
-      }
-    | {
-        __typename: 'JobMetadataEntry';
-        jobName: string;
-        repositoryName: string | null;
-        locationName: string;
-        label: string;
-        description: string | null;
-      }
-    | {
-        __typename: 'JsonMetadataEntry';
-        jsonString: string;
-        label: string;
-        description: string | null;
-      }
-    | {
-        __typename: 'MarkdownMetadataEntry';
-        mdStr: string;
-        label: string;
-        description: string | null;
-      }
-    | {__typename: 'NotebookMetadataEntry'; path: string; label: string; description: string | null}
-    | {__typename: 'NullMetadataEntry'; label: string; description: string | null}
-    | {__typename: 'PathMetadataEntry'; path: string; label: string; description: string | null}
-    | {
-        __typename: 'PipelineRunMetadataEntry';
-        runId: string;
-        label: string;
-        description: string | null;
-      }
-    | {__typename: 'PoolMetadataEntry'; pool: string; label: string; description: string | null}
-    | {
-        __typename: 'PythonArtifactMetadataEntry';
-        module: string;
-        name: string;
-        label: string;
-        description: string | null;
-      }
-    | {
-        __typename: 'TableColumnLineageMetadataEntry';
-        label: string;
-        description: string | null;
-        lineage: Array<{
-          __typename: 'TableColumnLineageEntry';
-          columnName: string;
-          columnDeps: Array<{
-            __typename: 'TableColumnDep';
-            columnName: string;
-            assetKey: {__typename: 'AssetKey'; path: Array<string>};
-          }>;
-        }>;
-      }
-    | {
-        __typename: 'TableMetadataEntry';
-        label: string;
-        description: string | null;
-        table: {
-          __typename: 'Table';
-          records: Array<string>;
-          schema: {
-            __typename: 'TableSchema';
-            columns: Array<{
-              __typename: 'TableColumn';
-              name: string;
-              description: string | null;
-              type: string;
-              tags: Array<{__typename: 'DefinitionTag'; key: string; value: string}>;
-              constraints: {
-                __typename: 'TableColumnConstraints';
-                nullable: boolean;
-                unique: boolean;
-                other: Array<string>;
-              };
-            }>;
-            constraints: {__typename: 'TableConstraints'; other: Array<string>} | null;
-          };
-        };
-      }
-    | {
-        __typename: 'TableSchemaMetadataEntry';
-        label: string;
-        description: string | null;
-        schema: {
-          __typename: 'TableSchema';
-          columns: Array<{
-            __typename: 'TableColumn';
-            name: string;
-            description: string | null;
-            type: string;
-            tags: Array<{__typename: 'DefinitionTag'; key: string; value: string}>;
-            constraints: {
-              __typename: 'TableColumnConstraints';
-              nullable: boolean;
-              unique: boolean;
-              other: Array<string>;
-            };
-          }>;
-          constraints: {__typename: 'TableConstraints'; other: Array<string>} | null;
-        };
-      }
-    | {__typename: 'TextMetadataEntry'; text: string; label: string; description: string | null}
-    | {
-        __typename: 'TimestampMetadataEntry';
-        timestamp: number;
-        label: string;
-        description: string | null;
-      }
-    | {__typename: 'UrlMetadataEntry'; url: string; label: string; description: string | null}
-  >;
-  assetKey: {__typename: 'AssetKey'; path: Array<string>} | null;
-};
-
 export type RunDagsterRunEventFragment_HookCompletedEvent = {
   __typename: 'HookCompletedEvent';
   message: string;
@@ -3454,7 +3293,6 @@ export type RunDagsterRunEventFragment =
   | RunDagsterRunEventFragment_ExecutionStepUpForRetryEvent
   | RunDagsterRunEventFragment_FailedToMaterializeEvent
   | RunDagsterRunEventFragment_HandledOutputEvent
-  | RunDagsterRunEventFragment_HealthChangedEvent
   | RunDagsterRunEventFragment_HookCompletedEvent
   | RunDagsterRunEventFragment_HookErroredEvent
   | RunDagsterRunEventFragment_HookSkippedEvent
