+++ python_modules/dagster-graphql/dagster_graphql/schema/logs/events.py
@@ -420,22 +420,6 @@ GrapheneAssetMaterializationFailureReason = graphene.Enum.from_enum(
 )
 
 
-class GrapheneHealthChangedEvent(graphene.ObjectType, AssetEventMixin):
-    class Meta:
-        interfaces = (GrapheneMessageEvent, GrapheneStepEvent, GrapheneDisplayableEvent)
-        name = "HealthChangedEvent"
-
-    def __init__(self, event: EventLogEntry):
-        dagster_event = check.not_none(event.dagster_event)
-        self.asset_health_changed = dagster_event.asset_health_changed_data
-        super().__init__(**_construct_asset_event_metadata_params(event, self.asset_health_changed))
-        AssetEventMixin.__init__(
-            self,
-            event=event,
-            metadata=self.asset_health_changed,
-        )
-
-
 class GrapheneFailedToMaterializeEvent(graphene.ObjectType, AssetEventMixin):
     class Meta:
         interfaces = (GrapheneMessageEvent, GrapheneStepEvent, GrapheneDisplayableEvent)
@@ -648,7 +632,6 @@ class GrapheneDagsterRunEvent(graphene.Union):
             GrapheneExecutionStepSuccessEvent,
             GrapheneExecutionStepUpForRetryEvent,
             GrapheneExecutionStepRestartEvent,
-            GrapheneHealthChangedEvent,
             GrapheneLogMessageEvent,
             GrapheneResourceInitFailureEvent,
             GrapheneResourceInitStartedEvent,
